Make the Choreographer frame delay configurable with a system property
The default 1ms frame delay is overkill, it causes an unnecessarily large
number of animation frames. Add an option to configure the default frame delay.
This is part of a larger effort to increaser the frame delay to something that
approximates 60fps. The benefit of this is that animations will require much
fewer frames, which will result in better performance.

PiperOrigin-RevId: 725688137