Set the default Choreographer frame delay in Robolectric to 15 ms (60 fps)
Previously, the default Choreographer frame delay was 1ms, which approximated
1000fps. This was overkill and caused an excessive number of animation frames
to be run, which caused performance issues when running tests that contained
long animations. Reduce the default frame delay to 15ms, which approximates
60ps.  This only applies to paused looper mode. Also, this value is still
configurable by a system property, since some tests depend on the 1ms value.

Note if your test fails as a result of this CL, just add:

    jvm_flags = ["-Drobolectric.defaultFrameDelayMs=1"]

to your android_local_test target(s) that are failing.

PiperOrigin-RevId: 727008546