Update a few Shadows for new Baklava.
For ShadowLauncherAppsTest, the signature changed so it needed another if
statement.
For ShadowOverlayManager, the signature changed.
For ShadowUiModeManager, the sGlobals object was being set to null, but
it doesn't always created every test. One test was recreating an activity
that was created in a previous test (on purpose), so sGlobals wouldn't be
recreated but it would still try to remove listeners from it. This
change made it so that it just resets globals rather than setting it to null.
This is more correct since the service it's based on doesn't actually
get deleted so this shouldn't either.

PiperOrigin-RevId: 729252695