Remove deprecated ProxyMaker and directlyOn proxy variant
The directlyOn proxy variant has been replaced by reflector long ago due to
reflector being more targeted and efficient.

Robolectric itself does not use the directlyOn proxy variant any more. It is
possible that third-party tests and third-party custom shadows are using it,
but it is fairly trivial to either migrate to the directlyOn reflection
variants or reflector.

PiperOrigin-RevId: 730613666