Add custom instrumentation configuration for the Simulator
When an Android app is built, desugaring process may update Kotlin stdlib
classes to reference desugared jdk libs. This requires that these transformed
Kotlin classes be acquired by the Robolectric class loader. Also, the default
Robolectric AndroidConfigurer is highly tuned to test libraries, which will not
be present when the simulator runs.

Add some custom instrumentation configuration for the simulator to ensure that
modified Kotlin classes can be acquired.

PiperOrigin-RevId: 731461698