Install a fake MediaProvider in Robolectric for APIs >= Q.
The Android platform provides a facility for apps to contribute and access media data that is
available on a device’s external storage volume. See
[shared media docs](https://developer.android.com/training/data-storage/shared/media).

Apps access the media store via the ContentProvider API. The AOSP implementation of this
ContentProvider is provided in a separate built-in system application, along with the system
photo picker and other associated facilities.

Because the implementation is in a separate apk, it is not available in Robolectric environments.
Attempting to resolve the MediaStore in Robolectric environments can result in errors like
`No content provider: content://media/external_primary/video/media/1` or otherwise incorrect
behavior.

This commit adds an initial fake implementation of the media provider when simulating Android
versions Q and up.

If this change causes your test to fail, you can revert back to the old behavior by passing this
jvm flag "-Drobolectric.installFakeMediaProvider=false"

PiperOrigin-RevId: 731831960