Set the context class loader when the simulator starts
The context class loader is used as the default class loader when
ServiceLoader.load is called. Without setting the context class loader,
services will be attempted to be loaded by the app class loader, which will not
have visibility into classes in the Robolectric class loaders. This fixes an
issue loading AndroidX Test plugins in the Simulator.

PiperOrigin-RevId: 734228269