Inject a Motion event at simulator startup
For some currently-unknown reason, the simulator may occasionally start with a
glitchy blank screen. This can be resolved manually by clicking on the screen.
Inject a fake motion event to trigger this behavior. At some point we may want to
investigate further to determine if something more specific can be done, such
as invalidating some areas or giving focus to some View.

PiperOrigin-RevId: 738415808