Add a `Simulator(Class activityClass)` constructor
This constructor takes an Activity class and initializes the Activity when when
the simulator starts. The goal is to ensure that the simulator setup logic is
invoked before the target Activity is created so that animations that are
created and started by the Activity are initialized and run after the call to
`ShadowChoreographer.setPaused(true);`. Otherwise, the animations will appear
to end immediately.

PiperOrigin-RevId: 743685363