Use fully-qualified RuntimeEnvironment class name in ShadowProviderGenerator
It is possible that shadow packages are not in the `org.robolectric`
package. For instance, users can use the annotation param:
-Aorg.robolectric.annotation.processing.shadowPackage=my.package

And the shadow package will be `my.package`.

Because of this, the it is much simpler to use the full-qualified
RuntimeEnvironment class name.

Also add an integration test that uses RAP to capture this issue.