Remove deprecated `ShadowApplication#getAppWidgetManager()` method
This commit removes the `ShadowApplication#getAppWidgetManager()` deprecated method.

It was deprecated in https://github.com/robolectric/robolectric/commit/98cc2e81f27e5e3e49f94a7519db536e250eab6b,
released with Robolectric 4.1.
It is no longer used in the project, and the migration is straightforward.

It also restores the test that was removed in the mentioned commit.