Add a highly experimental gradle plugin to run the Robolectric simulator
When included in an Android Gradle project, this plugin adds a task,
`simulate`, that will launch the Robolectric simulator. It can be
run using `./gradlew simulate`.

Also, extract some functionality to set POM metadata and sonatype repositories
into Kotlin extension methods to keep the build script DRY.

Unfortunately DeployedRoboJavaModulePlugin is not usable for a gradle plugin --
it is designed for a Java library.

PiperOrigin-RevId: 754704437