Add an option to use the member HardwareRenderer in ViewRootImpl
Certain Drawables, such as RippleDrawables, use RenderNodeAnimator, which
invokes HardwareRenderer.nRegisterAnimatingRenderNode on the HardwareRenderer
objects that are contained in View's attachInfo. In order to better support
RippleDrawables in RNG (and RenderNodeAnimator in general), the
HardwareRenderer associated with the ViewRootImpl should be used when
rendering.

This issue causes log noise with messages such as `Trying to start new
animators on 0x7f977daf7270 (RippleDrawable) without an animation handle!`

Add an option to use the member HardwareRenderer object. The default will be
flipped to true in a forthcoming CL. There is still a bug with the end
animations for ripple drawables. Once that's fixed the default will be true.

PiperOrigin-RevId: 755017118