Allow ShadowAccessibilityInputConnections to share state
An AccessibilityInputConnection is constructed each time InputMethod.getCurrentConnection() is called, but each of these
individual objects shares a connection via the internal
RemoteAccessibilityInputConnection to the Android system server.
This means that in real use, two AccessibilityInputConnections
can influence the state of each other: one could set a selection
and the other would read the new selection. This change ensures
that the same behavior is present in Robolectric by moving the
collection of shadowed methods to a
ShadowRemoteAccessibilityInputConnection. This has the added benefit
of using only real methods in ShadowAccessibilityInputConnection.

PiperOrigin-RevId: 755844468