ctesque: Only run UiAutomationTest on Emmulator before Android 15
UiAutomation#setRotation will set the rotation result to Settings.System, and will trigger
Android's updating for display in server part. But its updating will be overridden by the default
display rotation when there is deferring occurs(based on debugging). Before we find a proper
solution to fix it or avoid it, related tests are suppressed from Android 15.

Signed-off-by: utzcoz <utzcoz@outlook.com>