Disable testMatch on 35-36 for ShadowNativeColorSpaceTest.
ColorSpace.match() is a flaky function in real android. It attempts
to find a ColorSpace that matches the given Transform and TransformParameters.
It goes through the list of existing ColorSpaces which is a map from
a name to the ColorSpaces.

It iterates through the values() of that map to test for which is the
first one match the given values. Unfortunately, a number of the
ColorSpaces in the values will match each other with their
compare functions, so there's multiple correct answers in the list.
Since the list is random because it's from a map, this means the
test fails about half the time.

To fix this, we disabled the tests on 35 and 36 and just left main as this
bug will be fixed.

PiperOrigin-RevId: 759764096