Add functionality to block shadow BluetoothSocket connections until other events.
Real BluetoothSocket connections are blocking but not necessarily immediate operations.
The methods added here help exercise scenarios where the connection is blocked.
When the shadow blocks connect(), the connection would unblock either when a shadow-specific unblock method is called,
OR when the real close() method is called.
Per the connect() documentation, "close() can be used to abort this call from another thread."
If the shadow-specific unblock method is called, connect() will succeed.
If close() is called, connect() will fail (throwing an IOException).

PiperOrigin-RevId: 762092408