Update the SimulatorPlugin for publishing in the Gradle Plugin Portal
Gradle plugins are designed to be published in the Gradle Plugin Portal.
when a Gradle plugin is defined, a meta-data only publication artifact
is created that will not pass validation to be published in Maven
Central. Add the metadata required for the Gradle Plugin Portal.

Also, update the `simulate` task to automatically add a dependency on
the simulator artifact when it is invoked.