Set the simulator-gradle-plugin version to 1.0-SNAPSHOT
This will decouple the versioning of the simulator-gradle-plugin from
Robolectric. It will allow the plugin to be updated at a different
cadence. The plugin has to be published in the Gradle Plugin Portal, and
cannot be published in MavenCentral.  Long-term we may want to split out
the plugin into a separate git repo, but for now we can keep it in the
Robolectric repo.