Fix ApplicationInfo inconsistencies in Robolectric.
Previously Robolectric would separately create
and store an ApplicationInfo object in ShadowActivityThread.

There was a few problems with this:
1. The initial ApplicationInfo object was not fully populated,
notably fields like uid
2. Context.getApplicationInfo returned a different result than
PackageManager.getApplicationInfo.
3. Context.getApplicationInfo returned the direct reference to
ApplicationInfo. So tests could errorneously modify the ApplicationInfo object,
but that might not change the result of PackageManager.getApplicationInfo.

This commit modifies Robolectric's app bringup logic to statically
create and store the ApplicationInfo in PackageManager at an early point.

PiperOrigin-RevId: 769208473