Add ShadowTextureView to noOp drawing with it.
TextureView uses OpenGL to do it's drawing, so even the native
code will return 0 when trying to use nCreateTextureLayer. This
causes a crash because the hardwareRenderer/deferredRenderer is
0 so the initialization crashes.

This change makes the getTextureLayer function of TextureView do nothing so
it will at least continue the test and draw everything else even
if this isn't able to.

PiperOrigin-RevId: 776611916