Fix segfault when rendering AnimatedVectorDrawables in Android O through Q
In Android O through Q, AnimatedVectorDrawables are backed by
com.android.internal.view.animation.NativeInterpolatorFactoryHelper, which is
responsible for creating the native interpolators that are used in Animated
vector drawables. Previously, trying to render an AnimatedVectorDrawable in Q
and below could results in a segfault due to attempts to reference the null
interpolators.

PiperOrigin-RevId: 778719032