Add `addedInSdk` attribute to `@ClassName`
This allows specifying the SDK version in which a class referenced by `@ClassName` was introduced.
This is helpful when bumping the min SDK version, since we can easily find the `@ClassName` that are no longer needed.

This commit only focuses on adding the argument, to confirm that this suggestion is valid.
Then I'll submit other PRs to add it to the various usages of `@ClassName`.