Derive orientation from screen size
If no orientation is provided in qualifiers and a screen width or height is
provided, then derive it based on the width and height that are declared. Previously
the explicitly declared width and height could end up inverted when reconciling with
the previous orientation.

Issue: https://github.com/robolectric/robolectric/issues/10428
PiperOrigin-RevId: 780124932