Add waitForService, isDeclared and waitForDeclaredService for ShadowServiceManager.
The method waitForService should return the requested service if available,
otherwise block the thread to wait until service is available.

The method isDeclared should return whether the specific service is declared.
In production, service may declared in VINTF manifest or somewhere,
setServiceDeclaredOrNot is added to provide a way to setup the declared status for
the shadow.

The method waitForDeclaredService should first check whether service is declared or
not, if the service is declared, waitForService will be called; otherwise, return
null.

PiperOrigin-RevId: 780364522