Let ShadowWifiManager return invalid network ID for addNetwork.
Add a new call for simulating an error in the addNetwork call, which
returns -1 when there is an error. This can fail for various reasons,
e.g. invalid network configuration, or third-party apps attempting to
do this while targeting certain SDK levels (less than Q), or when it
is disallowed by the user.

PiperOrigin-RevId: 782214969