Support maxBounds in WindowConfiguration.
This requires updating both DecieConfig and ShadowDisplayManagerGlobal.
ShadowDisplayManagerGlobal has a function called changeDisplay that needs
to be updated so that it updates the default display when changing
its size. This also forces an update on the width/height of that display
so it picks up the changes. To test it, DeviceConfigTest has some cases
that test multiple displays and changing displays. There's also now
a DisplayTest that tests getRealSize.

PiperOrigin-RevId: 785549085