Adapt to constructor API changes in Android.
The in-development SDK has modified two constructors:

1. Two new params were added to the WifiUsabilityStatsEntry constructor, causing tests accessing
this codepath to fail with NoSuchMethodException. This commit adjusts
WifiUsabilityStatsEntryBuilder to pass in placeholder values
for these new parameters. And also refactors the condition statements for readability: they are
now ordered by API level ascending.

2. A parameter was removed from BackMotionEvent constructor. This commit refactors the
BackMotionEvent creation logic to be strictly SDK based, and adds a new condition for post-Baklava
constructor.

PiperOrigin-RevId: 786666817