Make ShadowApplicationPackageManager use resolveTypeIfNeeded, not getType
When trying to match intent filters, real android uses
resolveTypeIfNeeded when getting the type during Activity resolution.
In Robolectric, it was using getType(). This meant that if the type
was null in the intent but it could still figure out the type,
it would not get the right type and then fail.

There's a common bug that happens in android where the intent
has the Type set first and then the data, which will erase the
Type that was just set. In the case we saw, it would find the
proper type on emulators, but robolectric would stick with the
null type there.

PiperOrigin-RevId: 786814979