Mark LEGACY LooperMode as unsupported in upcoming SDKs.
The Android Baklava release offers a new "concurrent" MessageQueue implementation.
This new implementation will require significant rework in the Looper and MessageQueue
shadow classes. It doesn't make sense to spend effort adjusting the long-deprecated
legacy LooperMode shadows to work with this new implementation.

This commit adds documentation and a runtime check that will fail a test that
uses the Legacy LooperMode on SDKs > Baklava.

PiperOrigin-RevId: 788449683