Copy and initialize hyphenation data in DefaultNativeRuntimeLoader
For Android V and above, the path to hyphenation data is passed to the native
runtime using the `ro.hyphen.data.dir` native Android system property.
Otherwise, it is set using the `hyphen.data.dir` java system property.

The `android.text.Hyphenator` static initializer is invoked in Android P and
above. It is not supported on O.

For #10402.

PiperOrigin-RevId: 790866814