Ensure all the idleFor logic runs on the Looper thread.
Currently idleFor will execute in the calling thread, and
potentially hop over to the Looper thread for each separate
idle operation.

This is a wasteful number of thread hops, which will only
get worse once get*ScheduledTime APIs are converted to use
TestLooperManager APIs.

This commit moves the idleFor logic to its own Runnable posted
on the Looper thread.

PiperOrigin-RevId: 793686965