Add a script to generate a legacy http client bundle
This script takes the path to org.apache.http.legacy.jar (generated from AOSP
sources), and creates a Maven bundle.  This is necessary now that we dropped
support for LOLLIPOP. The jar file is needed for shadows-httpclient tests.
There is no supported android-all jar that contains the legacy http client
implementation.