Make Scheduler#size as unsupported on SDKs > Baklava.
LooperMode.PAUSED historically had an implementation of the legacy
Robolectric Scheduler API to make it easier for users to migrate
to PAUSED LooperMode.

Scheduler includes a 'size' API which requires reaching into MessageQueue
internals and is not supported on the new Baklava+ concurrent MessageQueue.
This API feels quite unnecessary as Android public SDK APIs like those in
Handler could serve the same purpose.

This commit modifies the Scheduler#size implementation to throw if it is
used on Android SDKs > Baklava.

PiperOrigin-RevId: 796471845