Update Javadoc of {get,peek}NextStartedService to better reflect the actual implementation
`getNextStartedService()` and `peekNextStartedService()` retrieve the first
element from the list of started services. Since new services are appended to
the end of the list, these methods return the oldest intent, not the most
recent as the Javadoc previously stated.

This also makes their behavior clearer in contrast to the
getNextStartedActivity methods, which does return the most recent item.

PiperOrigin-RevId: 799099331