Prepare to turn off perfetto logging in Baklava+.
Currently in newer SDKs ViewRootImpl will initialize
ProtoLog, which causes a background thread to start and
causes a bunch of activity dealing with perfetto data sources.

All of this is unnecessary in host environments.

This commit shadows the same method as ravenwood to direct
ProtoLog to use a simple implementation that logs straight to logcat.

PiperOrigin-RevId: 802220236