Prevent error logs and unnecessary work when reading flags on Baklava.
AconfigPackage is a new API in Baklava that is used to read flag values.
The real Android implementation will read metadata from files stored
on the device, and log errors when the files cannot be found.
Since those files don't exist in Robolectric, this is all
unnecessary work.

This commit adds a stub ShadowAconfigPackage that just returns
the default stub values.

PiperOrigin-RevId: 804620090