Prepare for IBluetoothProfileCallback change.
ShadowBluetoothAdapterTest#getProfileProxy_serviceListener and related tests were timing out
when executed on the latest indevelopment SDK. The root cause is the indevelopment SDK has a
change in BluetoothAdapter where service connection callbacks are now received via a
IBluetoothProfileCallback.

This commit shadows the new getProfile method so Robolectric can simulate the
callback method being invoked.

This commit also cleans up ShadowBluetoothAdapterTest so it fast-fails instead
of timing out if service listener is not invoked.

PiperOrigin-RevId: 806001382