Update loading preference order for acquired classes in the simulator
Previously, the preference order when loading acquired classes in the simulator was:

1) The deploy jar and any extra jars passed as args to the simulator
2) The Android framework jars
3) Any jars on the runtime classpath

After this change, (1) and (2) are swapped, so the order is:
1) The Android framework jars
2) The deploy jar and any extra jars passed as args to the simulator
3) Any jars on the runtime classpath

This is to ensure that if the deploy jar inadvertently has copies of framework
classes, we should continue to load the framework versions.

Note that this may cause outdated framework classes to be preferred over the
contents of the deploy jars. So in the future we may need finer-grained rules.

PiperOrigin-RevId: 811497476