Add support for Settings.{Global,Secure,System}.{get,put}StringForUser
Although these APIs are not public and are not included in the public stubs
jar, they can be used by system apps in the Android platform.

In Android's Settings class, the getIntForUser and putIntForUser methods just
call into the getStringForUser and putStringForUser variants, so ideally only
the string variants need to be shadowed. The existing getIntForUser and
putIntForUser shadow methods can be removed. Longer-term, the getStringForUser
and putStringForUser methods can support the majority of the Settings methods.

PiperOrigin-RevId: 819190653