Remove unnecessary shadow methods in ShadowSettings
This defers more of the Android Settings class in Robolectric to the real
framework methods. In real Android, all settings are converted to and stored as
strings. Robolectric had a custom implementation that stored settings with the
type information, which causes behavior differences. For example, in real
Android, if Settings.Global.putFloat is called, it can also be queried with
Settings.Global.getString. This previously did not work in Robolectric.

PiperOrigin-RevId: 823586928