Stop shadowing Settings.Secure.getInt methods
In real Android, calling Settings.Secure.getInt just invokes the getString
variants. The real Android framework code of Settings.Secure.getInt can be used
instead of shadowing and using a custom implementation. This requires changing
the underlying storage in ShadowSettings.Secure to store values as strings.
This also revealed an edge case when querying the location mode in Android P.
In Android P, the location mode from Settings.Secure is considered 'off' when
both GPS and network location providers are disabled, and Robolectric was not
following this behavior.

PiperOrigin-RevId: 829057967