Add support for asBinder in Robolectric system service interfaces
In Robolectric's ServiceManager, system service interfaces are mostly implemented
as null proxies - proxy objects that return null or zero for all methods. However,
some framework code, such as Car.java, calls the `asBinder` method on the system
service interface to check if the binder objects are the same. Add basic support
for this in Robolectric.

PiperOrigin-RevId: 830925369