Add integration test for junit vintage engine
The junit vintage engine is JUnit platform's official test engine
to support JUnit3 and JUnit4 tests. Robolectric extends from
JUnit4, so it's expected that we can run Robolectric test and
JUnit jupiter tests in a module when developers enable JUnit jupiter
and vintage engine for its test module. The new module
is used to verify this behavior.

Signed-off-by: utzcoz <utzcoz@outlook.com>