Use more code from Android in qualifier parsing.
Previously Robolectric had custom logic that
converted a Configuration+DisplayMetrics to and from a qualifier String.
The code to parse a qualifier String was especially undesirable as it
depending on the binary resources java implementation.

This commit converts this logic to use Configuration.resourceQualifierString
to produce a qualifier String where applicable.

There isn't a java logic available in the Android framework to parse
a qualifier String. But this commit simplifies Robolectric's existing
logic to parse the String directly into a Configuration object,
and thus can use Configuration.updateFrom to copy it.

Included in these changes is that the qualifier String will now
contain DisplayMetrics pixel data where applicable.

PiperOrigin-RevId: 839482811