Allow specifying a display pixel width/height in qualifiers.
Historically Robolectric only allowed specifying a screen size in
density-independent pixel (dp) units. This was restricting because it
was difficult to accurately specify and represent real physical devices
whose attributes are typically specified in pixels.

For backwards compatibility, this change will make Robolectric's qualifier
parser use the width/height  in pixels iff it is specified and dp units
are not.

PiperOrigin-RevId: 839486408