Future proof IBluetoothManager support.
IBluetoothProfileServiceConnection is being deleted in the current version of
Android. This causes build and runtime breakages in
IBluetoothManagerDelegates which has a direct reference to this class.

This commit refactors IBluetoothManagerDelegates to use a shadow of
IBluetoothManager.Default on newer SDKs where it is available. Doing so
facilitates using @ClassName to refer to IBluetoothProfileServiceConnection
by reflection on older SDKs.

PiperOrigin-RevId: 839780358