Fix ShadowCaptureRequestBuilderTest in Gradle
For Android SDK levels B and below, there is an Android framework bug in the
reflection logic in CameraMetadata.getKeys, in which the __robo_data__ fields
get included in the field iteration. This will be fixed in the upcoming Android
release. Add a temporary workaround by shadowing CameraMetadata.getKeys. Once
the preinstrumented jars are updated, this shadow can be removed.

PiperOrigin-RevId: 842789433