Simplify the perf stats Metadata API.
The PerfStatsCollector Metadata class currently has
a complex API that allows storing arbitrary classes.

However, this capability is unused. The only data
that is stored in the Metadata is the SDK int.

This commit simplifies the API by removing the AndroidMetadata
class, and just storing the SDK int directly in Metadata class.

This resolves the classloader issues that can occur when
org.robolectric.AndroidMetadata is stored and retrieved
by different class loaders/

PiperOrigin-RevId: 845346598